fix copy and paste error
