+++ services/QuillDiagnostic/package.json
@@ -127,7 +127,7 @@
     "pos": "^0.4.2",
     "promise-polyfill": "^6.0.2",
     "pusher-js": "^4.1.0",
-    "quill-component-library": "^0.1.21",
+    "quill-component-library": "^0.1.29",
     "quill-marking-logic": "^0.6.1",
     "raven-js": "^3.17.0",
     "react": "15.6.2",
@@ -144,7 +144,7 @@
     "react-select-search": "0.5.5",
     "react-selectize": "^2.1.0",
     "react-simple-pie-chart": "^0.4.1",
-    "react-sortable": "^1.2.0",
+    "react-sortable": "^2.0.0",
     "react-transition-group": "^2.2.1",
     "react-wakelock": "0.0.4",
     "redux-persist": "^4.8.1",
